refactor: createMetaBox
